requirements.txt
torch>=1.7.0
torchvision>=0.8.1
numpy>=1.19.2
Pillow>=8.0.1
tqdm>=4.51.0
matplotlib>=3.3.3
opencv-python>=4.4.0.46
tensorboard>=2.4.0
PyYAML>=5.3.1
scipy>=1.5.4
pandas>=1.1.4
ftfy>=6.0.3
indic-nlp-library>=0.71
indicnlp>=0.1
tensorflow>=2.4.0